Rework a bit
