--- samples/multi-module-compose/README.md
@@ -1,10 +1,0 @@
-# Module Structure
-
-```mermaid
-flowchart TD
-    app --> screen-details
-    app --> screen-home
-    app --> core
-    screen-details --> core
-    screen-home --> core
-```
