+++ samples/multi-module-viewmodels/core/src/commonMain/kotlin/dev/zacsweers/metro/sample/multimodule/viewmodels/core/StaticCompositions.kt
@@ -1,0 +1,8 @@
+// Copyright (C) 2025 Zac Sweers
+// SPDX-License-Identifier: Apache-2.0
+package dev.zacsweers.metro.sample.multimodule.viewmodels.core
+
+import androidx.compose.runtime.staticCompositionLocalOf
+
+val LocalViewModelGraphProvider =
+  staticCompositionLocalOf<ViewModelGraphProvider> { error("No ViewModelGraphProvider registered") }
